converted tabular data to counted values.
At first when we loade data throgh .csv file it is in posibilities form like 1s and 0s.
So to work with them we have to count them accordingly to calculate probabilites.
This task is done

Futur work:
it include all the inilitial probabilies which is always going to be calculated like
P(D1) D1 meaans here disease1 in the data and same for all the other diseases.